got the index creation parallelized as well
the readdirs aren't.  they just go serially and make a list
of every index file

then the opening and reading of those index files is threaded

one funny by-product of this is that we've weakened further the
semantic of how we handle overlapped writes.  it's not even
deterministic anymore since the order in which multiple index
files is read is not deterministic.

if the same pid does the overwrite than it will be fine.
otherwise, all bets are off.

needs a lot of testing



git-svn-id: https://plfs.svn.sourceforge.net/svnroot/plfs/trunk@189 979fb63a-54b4-4514-8d67-9ab8ea1b59df