fuse/plfs_fuse.cpp
Added if statement around code that was using getpwid(uid) to grab 
the user database entry for uid. Posix testing was using a uid that did not
contain an entry in the user datebase so this code would cause a segfault 
in this situation.

src/plfs.cpp

In the plfs_open funciton, if the writefile was not being opened correctly 
the writefile would be set to NULL. A later piece of code was then trying to
get the reference count from NULL which was causing problems.

 


git-svn-id: https://plfs.svn.sourceforge.net/svnroot/plfs/trunk@241 979fb63a-54b4-4514-8d67-9ab8ea1b59df