removed a race condition from is_plfs_file:
From the email explaining it to Gary:

Yep.  Silly race condition and returning a binary value when we
should have been returning more info.

getattr needs to know if something is:
a container, a regular file, a directory, a symlink, non-existent

we had code that basically said:
if F is a container, do plfs container stat on F
else do normal stat on F   

but the code needs to say:
if F doesn't exist, return ENOENT
else if F is a container, do plfs stat on F
else do normal stat on F

here's what was happening:

A: is F a container?  No, it's not a container (bec doesn't exist)   
B: mkdir F.tmp, create F.tmp/access, rename F.tmp F
A: ok, it's not a container, I'll return a normal stat of it
A: the normal stat of it says it's a directory (WHOOPS!)

Here's the basic fix:

getattr(F):
  switch F:
    case ENOENT: return ENOENT;
    case container: return plfs_getattr(F);
    default: return lstat(F)

The other changes are to change signatures since is_plfs_file now takes an 
extra arg to return something else by value



git-svn-id: https://plfs.svn.sourceforge.net/svnroot/plfs/trunk@288 979fb63a-54b4-4514-8d67-9ab8ea1b59df