fixed Gary's permission bug
Gary's code was calling open with O_CREAT but no mode

when this goes to PanFS, it makes a file with junk bits:
---sr----T

when this was going to PLFS, it was calling dirMode(---sr----T)
and creating the topLevel like that which then didn't allow
the dir to be read.  Gary was also opening with O_RDWR and
the create of the index was failing because the dir didn't
allow the user to read things in there.

We weren't adding S_IRUSR in dirMode() which meant that
users couldn't create write-only files before.  They still
probably can't since PLFS won't protect users' files from
themselves but at least the stats look correct.

So the problem was:

1) gary was opening w/ O_RDWR and O_CREAT and no mode
2) plfs wasn't adding S_IRUSR in the dirMode()



git-svn-id: https://plfs.svn.sourceforge.net/svnroot/plfs/trunk@295 979fb63a-54b4-4514-8d67-9ab8ea1b59df