ok, I fixed the readdir on the root directory but I'm scared about it since ./configure of plfs on plfs works so horribly right now on my mac.  it looks like the problem is partial overwrites that happen when I open the tarball.
so this shouldn't be related but it's still worrisome

Adam, maybe you can check this out early?  Maybe if it's OK, we can get this in
to the DST still?  It's a trivial amount of code and it's only used in just
plfs_readdir so I'm sure it's fine but boy do we need to fix the overlapped
writes thing if we want to run on mac's because it's really bad right now.



git-svn-id: https://plfs.svn.sourceforge.net/svnroot/plfs/trunk@343 979fb63a-54b4-4514-8d67-9ab8ea1b59df