rev'ing to 1.1.8rc
From the ChangeLog:

* Fixed the getattr bug (we were just getting the attr from the openfile
handle and not getting it from the Container as well).  This made the size
reasonable but the permissions/ownership/everthing else were garbage).
* fixed ADIO parse conf error.  Now a good error is reported when the 
conf file is bad or the file path is bad.
* added global_summary_dir to the plfsrc
* fixed tools/plfs_flatten_index which broke due to the stop_buffering
thing we put into the Index to stop buffering when memory is exhausted.
the buffering is so everyone retains their write index in memory to do
the flatten on close in ADIO



git-svn-id: https://plfs.svn.sourceforge.net/svnroot/plfs/trunk@456 979fb63a-54b4-4514-8d67-9ab8ea1b59df