was another weird problem with truncate
oh, truncate on an open file was using like lazy stat (sz_only) to 
plfs_getattr which was therefore only querying the open file (i.e.
not reading the individual index files) and the open file was a
newly opened file and it thought the size was 0 so truncate never
thought it had any work to do.



git-svn-id: https://plfs.svn.sourceforge.net/svnroot/plfs/trunk@580 979fb63a-54b4-4514-8d67-9ab8ea1b59df