But anyway.  I'm now created a reproducer for the access issue that the code team was dealing with where they misuse PLFS to ask questions about non-PLFS files. I put it into tests/access.C.  There's some internal documentation but basically you just build it and run ./access -t /mnt/plfs/some/path/to/plfs/file and access will walk that tree calling plfs_access on each component.  Currently it fails with plfs_access /mnt: ENOENT bec I wrote the reproducer but haven't yet coded up the fix.
Also, to build this, you'll need to first rerun autogen.sh since I had to edit
a Makefile.am in order to get the build system  to build it.  Also, when you
build it, you'll need to have a plfs library installed somewhere and you'll
need the build to find it.  I've achieved this by adding the path to my plfs
installation to:

CPLUS_INCLUDE_PATH
DYLD_LIBRARY_PATH



git-svn-id: https://plfs.svn.sourceforge.net/svnroot/plfs/trunk@645 979fb63a-54b4-4514-8d67-9ab8ea1b59df