allow named facilities to have both an abbreviated name and a long name (which is optional).  the long name, if used, could be more "human readable" ...
we'll use the long name (if present) when reporting the mask levels
(using the mlog_getmasks() API) and we'll take either name in 
mlog_setmasks().   the name facility functions get a 2nd name arg.

also, convert the mask functions to be case insensitive by using
strcasecmp()/strncasecmp().   the SUS standard says this requires
<strings.h>, so we include that in addition to <string.h>.  (we can
revert back to strcmp if this is an issue...).


git-svn-id: https://plfs.svn.sourceforge.net/svnroot/plfs/trunk@737 979fb63a-54b4-4514-8d67-9ab8ea1b59df