A branch with a HUGE amount of change from trunk 809.
This includes predominantly the new flat_file code which also
required a big rewrite to put container mode and flat_file mode
into instantiations of a virtual parent class.  This work is 
great bec it makes the code much easier to read and much cleaner.

This also includes the code that made the SC11 demo work which is
an important step forwards towards burst buffer.  This includes a
new plfsrc directive shadow_backends which forces compute nodes to
use a specified subset of the backends for their physical storage.



git-svn-id: https://plfs.svn.sourceforge.net/svnroot/plfs/branches/flat_file2@811 979fb63a-54b4-4514-8d67-9ab8ea1b59df