Initial changes that attempt to get PLFS to work with data-sieving
There is a new file, ad_plfs_features.c, that puts together a function
that returns some capabilities for PLFS. Makefiles had to be updated to
include this new file. A member in the struct in ad_plfs.c was changed
to use this new function in ad_plfs_features.c.

This new file should mitigate the need for the bug-fixes in the prep
patches that were there to attempt to make sure files were not opened
read-write.