Bring in a newer version of FindJNI.cmake.   The old one in Ubuntu10 didn't support JNI_FOUND.
PVFS appears to decide to link to OpenSSL at configuration time
if it finds it, otherwise it just compiles without it.  That means
you need to link to OpenSSL if PVFS did or you'll get unresolved
symbols at link time.   Put in some cheesey cmake hooks to help
with this.