First pass at bringing plfs.spec up to 2.4.
An rpm can be built with this spec file. However, we still need to
decide on what to do about VERSION and VERSION.LAYOUT. The spec file
currently has lines that work with these files commented out as cmake
doesn't do anything with them during a 'make install'. A changelog also
needs to be added.