fixed a performance issue where if the user did plfs_open with both O_CREAT and O_TRUNC, the container_open called both container_create, which did a truncate, and it also called container_truncate, which did it again
One fix would have been to not truncate in container_create and only
truncate in container_truncate but it's a bit more efficient to do
the truncate in container_create when possible.

So we fixed this by just not calling the container_truncate when we
realize the the container_create already did the truncate for us.