Update where rc1 tag is and the name of make dist's tarball
The rc1 tag is moved to the 5 so that resulting strings are 2.5rc1
rather than 2.5.0rc1. This makes the naming consistent with previous
PLFS versions.

The name of the generated tarball now uses plfs_VERSION_FULL instead of
coming up with its own name. Now the name of the tarball will be
consistent with all other version strings (ie the version string that
ends up in the library name).