Update spec file and documentation on how to test it.
The spec file can now handle version strings of variable length without
having to change the spec file beyond the version numbers. This means
that it isn't necessary to change what files need to be included when
the version string changes a library name beyond what the rpm syntax can
handle. For example, the spec file can deal with 2.5.1 just as easily as
it deals with 2.5; only the version numbers have to be updated.