add +ContainerEntry::older_than() to allow for for tie breaking when timestamps match.
don't back up "prev" in a map if it is already at map.begin()
(behavior not defined in C++, best not to do it).

if we have an overlap and the entry we are trying to insert
is size 0, then we can just discard it rather than trying to
run it through the overlap handler.

fix minor error in setting "first" in the overlap handler.