to be safe, check for any remaining open data droppings when doing a final close of a Plfs_fd.   close any found.
XXX: currently this can happen with PLFS/FUSE if a process
opens a file and then forks off a child to do the write.
 the forked proc doesn't generate an open, but it gets sent
to plfs_write() and makes a dropping.  e.g.

	# (eval "gcc -v >&1") 2>conftest.err

even if we fix this (e.g. to have writes use the pid of the opener
rather than the writer), we should keep this code anyway just to
play it safe (doesn't cost much to check).